Update app structure to components/hooks/utils Move old files to old folder and do not lint them
