+++ src/app/old/Device/Device.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
-import { fetchDataObj } from '@app/utils/commonFunctions';
-import { enableRCAgent } from '@app/utils/commonFunctions';
-import { device } from '@app/utils/commonDataTypes';
+import { fetchDataObj } from '@app/old/utils/commonFunctions';
+import { enableRCAgent } from '@app/old/utils/commonFunctions';
+import { device } from '@app/old/utils/commonDataTypes';
 import { useAuth } from 'react-oidc-context';
 import { LogViewer } from '@patternfly/react-log-viewer';
-import { RemoteControl } from '@app/Device/rc';
+import { RemoteControl } from '@app/old/Device/rc';
 import YAML from 'yaml';
 import {
     Accordion,
