+++ src/app/old/ImageBuilder/ImageBuilder.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { fetchData } from '@app/utils/commonFunctions';
+import { fetchData } from '@app/old/utils/commonFunctions';
 import { useAuth } from 'react-oidc-context';
 import {
   Card,
@@ -22,7 +22,7 @@ import { set } from 'yaml/dist/schema/yaml-1.1/set';
 
 
 
-const GitRepositories: React.FunctionComponent = () => {
+const ImageBuilder: React.FunctionComponent = () => {
   const [isLoading, setIsLoading] = React.useState(false);
   const auth = useAuth();
 
@@ -31,11 +31,11 @@ const GitRepositories: React.FunctionComponent = () => {
 
   return (
     <PageSection>
-      <Title headingLevel="h1" size="lg" style={{ marginBottom: '15px' }}>Git Repositories</Title>
+      <Title headingLevel="h1" size="lg" style={{ marginBottom: '15px' }}>ImageBuilder</Title>
 
     </PageSection>
   )
 };
 
-export { GitRepositories };
+export { ImageBuilder };
