+++ src/app/old/utils/commonFunctions.tsx
