+++ src/app/old/utils/fleetFormView.tsx
