+++ src/app/routes.tsx
@@ -1,21 +1,21 @@
 import * as React from 'react';
 import { Route, RouteComponentProps, Switch, useLocation } from 'react-router-dom';
-import { Overview } from '@app/Overview/Overview';
-import { EnrollmentRequests } from '@app/EnrollmentRequests/EnrollmentRequests';
-import { Fleet } from '@app/Fleet/Fleet';
-import { Devices } from '@app/Devices/Devices';
-import { Enroll } from '@app/Enroll/Enroll';
-import { Device } from '@app/Device/Device';
-import { GitRepositories } from '@app/GitRepositories/GitRepositories';
-import { Organization } from '@app/Organization/Organization';
-import { ImageBuilder } from '@app/ImageBuilder/ImageBuilder';
-import { NotFound } from '@app/NotFound/NotFound';
-import { useDocumentTitle } from '@app/utils/useDocumentTitle';
-import { RemoteControl } from '@app/Device/rc';
-import { Workload } from '@app/Workload/Workload';
-import { Experimental } from '@app/Experimental/Experimental';
-import { Experimental2 } from '@app/Experimental/Experimental2';
-import { Experimental3 } from '@app/Experimental/Experimental3';
+import { Overview } from '@app/old/Overview/Overview';
+import { EnrollmentRequests } from '@app/old/EnrollmentRequests/EnrollmentRequests';
+import { Fleet } from '@app/old/Fleet/Fleet';
+import { Devices } from '@app/old/Devices/Devices';
+import { Enroll } from '@app/old/Enroll/Enroll';
+import { Device } from '@app/old/Device/Device';
+import { GitRepositories } from '@app/old/GitRepositories/GitRepositories';
+import { Organization } from '@app/old/Organization/Organization';
+import { ImageBuilder } from '@app/old/ImageBuilder/ImageBuilder';
+import { NotFound } from '@app/old/NotFound/NotFound';
+import { useDocumentTitle } from '@app/old/utils/useDocumentTitle';
+import { RemoteControl } from '@app/old/Device/rc';
+import { Workload } from '@app/old/Workload/Workload';
+import { Experimental } from '@app/old/Experimental/Experimental';
+import { Experimental2 } from '@app/old/Experimental/Experimental2';
+import { Experimental3 } from '@app/old/Experimental/Experimental3';
 import FleetList from './components/Fleet/FleetList';
 import CreateFleet from './components/Fleet/CreateFleet';
 let routeFocusTimer: number;
