+++ src/index.tsx
@@ -17,7 +17,7 @@ declare global {
 const KEYCLOAK_AUTHORITY = process.env.NODE_ENV === 'production' ? window.env.REACT_APP_KEYCLOAK_AUTHORITY : process.env.REACT_APP_KEYCLOAK_AUTHORITY || "http://localhost:9080/realms/flightctl"
 const KEYCLOAK_CLIENTID = process.env.NODE_ENV === 'production' ? window.env.KEYCLOAK_CLIENTID : process.env.REACT_APP_KEYCLOAK_CLIENTID || "flightctl-ui"
 const KEYCLOAK_REDIRECT = process.env.NODE_ENV === 'production' ? window.env.KEYCLOAK_REDIRECT : process.env.REACT_APP_KEYCLOAK_REDIRECT || "http://localhost:9000" 
-const WS_URL = process.env.NODE_ENV === 'production' ? window.env.WS_URL : process.env.WS_URL || "http://localhost:8082" 
+// const WS_URL = process.env.NODE_ENV === 'production' ? window.env.WS_URL : process.env.WS_URL || "http://localhost:8082" 
 const oidcConfig = {
   authority: KEYCLOAK_AUTHORITY,
   client_id: KEYCLOAK_CLIENTID,
